requirements.txt
aiogram
instaloader
requests
python-dotenv
